requirements.txt
openai-whisper
sounddevice
keyboard
pyperclip
scipy
numpy
pystray
pillow
